Separated CartItems into its own component and included it on the PaymentDetails component upon a successful payment
