+++ lib/getItemInfo.js
@@ -1,0 +1,5 @@
+export default function getItemInfo(products, item) {
+    const itemInfo = products.filter(product => product.id === item.id)[0]
+    const priceInCents = itemInfo.price * 100 * item.quantity
+    return [itemInfo, priceInCents]
+}
