+++ src/app/api/payment/route.js
@@ -4,14 +4,16 @@ const stripe = new Stripe(process.env.STRIPE_SECRET_KEY)
 
 export async function POST(req) {
     try {
-        const body = await req.json()
-        const { amount } = body
+        const { amount, items } = await req.json()
 
         const paymentIntent = await stripe.paymentIntents.create({
             amount,
             currency: 'usd',
+            metadata: { items: JSON.stringify(items) },
             automatic_payment_methods: { enabled: true }
         })
+
+        //Code to empty the cart
 
         return new Response(JSON.stringify({ clientSecret: paymentIntent.client_secret }), {
             status: 200,
@@ -35,10 +37,7 @@ export async function GET(req) {
         return new Response(JSON.stringify({
             amount: paymentIntent.amount,
             status: paymentIntent.status,
-            items: [
-                { name: 'Test Product A', price: 1000 },
-                { name: 'Test Product B', price: 2000 }
-            ]
+            items: paymentIntent.metadata?.items ? JSON.parse(paymentIntent.metadata.items) : []
         }), {
             status: 200,
             headers: { 'Content-Type': 'application/json' }
