+++ src/app/payment-success/page.jsx
@@ -1,9 +1,12 @@
+import fetchProducts from "../../../lib/woocommerce"
 import PaymentDetails from "@/components/PaymentDetails"
 
-export default function PaymentSucess() {
-    return (<div>
-        <h1>Payment Successful! 🎉</h1>
+export default async function PaymentSuccess() {
+    const products = await fetchProducts()
+
+    return (<div className="container mx-auto p-6">
+        <h1 className="text-3xl font-bold mb-4">Payment Successful! 🎉</h1>
         <p>Thank you for your purchase.</p>
-        <PaymentDetails />
+        <PaymentDetails products={products} />
     </div>)
 }
