+++ src/components/Checkout.jsx
@@ -1,40 +1,16 @@
 "use client"
-import { useCart } from "@/app/context/CartContext";
-import Image from "next/image";
+import { useCart } from "@/app/context/CartContext"
 import { useState, useEffect } from "react";
 import { FaEdit } from "react-icons/fa"
 import getChangeQuantity from "../../lib/getChangeQuantity"
-import ChangeQuantitySpans from "./ChangeQuantitySpans"
+import calculateTotal from "../../lib/calculateTotal"
+import CartItems from "./CartItems"
 import StripeForm from "./StripeForm"
 import { Elements } from '@stripe/react-stripe-js'
 import { loadStripe } from '@stripe/stripe-js'
 
 export default function Checkout({ products }) {
     const { cartItems, addToCart, updateQuantity, removeFromCart } = useCart()
-
-    const getItemInfo = item => {
-        const itemInfo = products.filter(product => product.id === item.id)[0]
-        const priceInCents = itemInfo.price * 100 * item.quantity
-
-        return [itemInfo, priceInCents]
-    }
-
-    const calculateTotal = () => {
-        let initialTotal = 0
-
-        cartItems.map(item => {
-            const priceInCents = getItemInfo(item)[1]
-            initialTotal += priceInCents
-        })
-
-        return initialTotal
-    }
-
-    const formatPrice = priceInCents => {
-        const dollars = Math.floor(priceInCents / 100)
-        const cents = priceInCents % 100
-        return `$${dollars}.${cents.toString().padStart(2, '0')}`
-    }
 
     let changeQuantity = getChangeQuantity({ addToCart, removeFromCart, updateQuantity })
     changeQuantity.push({
@@ -45,11 +21,11 @@ export default function Checkout({ products }) {
         }
     })
 
-    const [total, setTotal] = useState(calculateTotal)
+    const [total, setTotal] = useState(calculateTotal(cartItems, products))
     const [editID, setEditID] = useState(null)
     const [newQty, setNewQty] = useState('')
 
-    useEffect(() => setTotal(calculateTotal()), [cartItems])
+    useEffect(() => setTotal(calculateTotal(cartItems, products)), [cartItems])
 
     const [clientSecret, setClientSecret] = useState('');
 
@@ -58,7 +34,10 @@ export default function Checkout({ products }) {
             fetch('/api/payment', {
                 method: 'POST',
                 headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({ amount: total })
+                body: JSON.stringify({
+                    amount: total,
+                    items: cartItems
+                })
             }).then(res => res.json())
                 .then(data => setClientSecret(data.clientSecret))
         }
@@ -74,64 +53,19 @@ export default function Checkout({ products }) {
             {clientSecret && (<Elements stripe={stripePromise} options={options}>
                 <StripeForm />
             </Elements>)}
-            <div>
-                <h2 className="font-bold text-xl mb-2">Order Summary</h2>
-                <ul>
-                    {cartItems.map(item => {
-                        const [itemInfo, priceInCents] = getItemInfo(item)
-                        const imageInfo = itemInfo.images[0]
-
-                        return (<li key={item.id} className="grid grid-cols-[100px_auto] gap-2 mb-2">
-                            <Image
-                                src={imageInfo.src}
-                                alt={imageInfo.alt}
-                                width={100}
-                                height={100}
-                                className="object-cover aspect-square"
-                            />
-                            <div className="text-sm">
-                                <p>{itemInfo.name}</p>
-                                <div className="grid grid-cols-2">
-                                    <div className="flex items-center flex-wrap gap-1">
-                                        {editID === item.id ? (<input
-                                            className="w-12 text-xs p-2 pr-0"
-                                            type="number"
-                                            min="0"
-                                            value={newQty}
-                                            autoFocus
-                                            onChange={e => setNewQty(e.target.value)}
-                                            onBlur={() => {
-                                                const qty = parseInt(newQty)
-                                                if (!isNaN(qty)) {
-                                                    if (qty == 0) removeFromCart(item.id)
-                                                    else if (qty > 0) updateQuantity(item.id, qty)
-                                                }
-                                                setEditID(null)
-                                            }}
-                                            onKeyDown={e => {
-                                                if (e.key === "Enter") e.target.blur()
-                                            }}
-                                        />) : (<>
-                                            {item.quantity > 1 && (
-                                                <span className="text-zinc-300 mr-1 align-middle whitespace-nowrap">
-                                                    x {item.quantity}
-                                                </span>
-                                            )}
-                                        </>)}
-                                        <ChangeQuantitySpans cqs={changeQuantity} item={item} />
-                                    </div>
-                                    <p className="text-right text-green-600">{formatPrice(priceInCents)}</p>
-                                </div>
-                            </div>
-                        </li>)
-                    })}
-                </ul>
-                <hr />
-                <div className="grid grid-cols-2 text-sm mt-2">
-                    <p className="font-bold">Total</p>
-                    <p className="text-right">{formatPrice(total)}</p>
-                </div>
-            </div>
+            <CartItems
+                cartItems={cartItems}
+                products={products}
+                total={total}
+                editable={true}
+                removeFromCart={removeFromCart}
+                updateQuantity={updateQuantity}
+                editID={editID}
+                setEditID={setEditID}
+                newQty={newQty}
+                setNewQty={setNewQty}
+                changeQuantity={changeQuantity}
+            />
         </div> : <p>Your cart is empty</p>}
     </>)
 }
